fix: функция changeLang пересена в index.js для устранения ошибок eslint
